Rename the files for clear format
