+++ src/pages/chat-room-list/App.js
@@ -1,6 +1,6 @@
 import React, { useState, useEffect, useRef, useCallback } from 'react';
 import CredentialsForm from '../../components/CredentialsForm';
-import HomeButton from '../../components/HomeButton';
+import HomeButton from '../../components/MenuButton';
 import ChatRoomList from './ChatRoomList';
 import styles from './submitForm.module.css';
 import * as Callbridge from '@iotum/callbridge-js';
