+++ src/pages/chat-room-list/ChatRoomList.js
@@ -1,7 +1,7 @@
 import React from 'react';
 import ChatRoom from './ChatRoom';
 import styles from './chat.module.css';
-import HomeButton from '../../components/HomeButton';
+import HomeButton from '../../components/MenuButton';
 
 const ChatRoomList = ({ rooms, onRoomClose }) => {
   return (
